Use Shortcut for key navigation of the main image view
The list view supports key navigation but only if it has focus. This
means that as soon as any tool button or other item is clicked, the
focus is lost and key navigation stops working. To prevent that, we use
Shortcut as a method that doesn't need focus.

Left/right will now always move through the images. Space will now stop
the slideshow if it is currently running.