Use a double-buffering like technique to only partially load large images
When dealing with very large images (currently defined as the largest
dimension being larger than 5000 pixels), loading the entire image into
memory makes memory usage spike to such an extent that we potentially
overload the system. This can be avoided by using Image's sourceSize and
sourceClipRect properties.

Unfortunately, with an async image, modifying these properties will
first clear the Image before loading the data again. To avoid this, this
implements a fake double-buffering approach where we have one active
image and one inactive image. The inactive image is loaded with new
data, then once it is done we swap these, making the inactive image
visible and hiding the old image.

Since we only have a partial view of the image now, a base layer is
added that is a scaled down version of the main image, which is
displayed when we don't have proper data for the visible area yet.