Add InfoSidebar, InfoDrawer and TagInput
InfoSidebar is a QQC2 Page that is meant to be separated from the
ImageViewPage's ListView by a moveable separator.

InfoDrawer is basically the same as the OverlayDrawer that was there
before.

InfoSidebar and InfoDrawer share the same content via
InfoDrawerSidebarBase.

TagInput is a QQC2 ComboBox that allows tags to be added to images.

ImageViewPage: Use InfoSidebar and InfoDrawer