Introduce FileInfo as a QML object to provide information about a file
This allows us to encapsulate some logic that is currently scattered
around different places and provides a central point for information
about files. It includes reading the width and height of images so that
we do not need to rely on Image's sourceSize for that.

It has been designed to be very lightweight to use from QML, to avoid
having to share instances and instead make it cheap to just create
multiple instances. Therefore, the actual information is cached in a
shared cache and information retrieval happens on a separate thread.