Support color correction of images
This makes use of the colorSpace support introduced in Qt 5.14 and 5.15
to do color correction on loaded images.

The color space support in Qt turns out to be a bit limited, as it does
not support LUT profiles and several other profiles also fail, but
images with "real" profiles (as opposed to those with extreme values for
testing) seem to work fine.

The display profile retrieval code matches that of Gwenview, except it
is using XCB rather than Xlib, as I personally found the XCB code a bit
more readable.