Only create a context drawer on mobile
On desktop, we never show the context drawer so it pointless to create
one and slows down startup unnecessarily.

The only effect this has on mobile is to change the context drawer title
of ImageViewPage back to the Kirigami default of "Actions". The "Edit
Image" that was set in the custom context drawer does not make sense
here anyway, since none of the actions in the drawer relate to image
editing.