Refactor and split up ImageDelegate
This changes ImageDelegate to BaseImageDelegate and splits out the four
different versions into different subitems, containing only the logic
that is relevant for that version of the delegate. To facilitate the
splitting, ImageViewPage is now fully responsible for deciding which
version of the delegate to show.

The logic around content types has also been cleaned up to make use of
the "type" property of FileInfo as a single source of truth for which
type of content is currently being viewed. This cleans up a bunch of
nasty duplicated logic in various property bindings that was hard to
follow and error-prone.