Push pages by URL, do not unnecessarily create components for them
Creating components means loading and parsing the full contents of
a page, including all its dependencies, this adds unnecessary
overhead during startup, especially for things that are rarely used
such as the about page. Instead just pass an URL of a page to load,
PageRow will then load and instantiate it when the page gets pushed.

The only caveat is that we do not know the URL of Kirigami's
AboutPage.qml, so instead, we create a local page that just forwards
to Kirigami AboutPage.