WIP
Trying to fix sidebar and thumbnail bar interaction — splitter position is not updated when info sidebar not visible, and info sidebar appears to animate width and splitter movement? closes faster than opens --- it should open at the side it wants offscreen and animate in to meet the splitter. Splitter mechanics should be changed on window resize too, and better min/max specified rather than using info implicit width as max which is variable.