ImageViewPage: Refactor resizable info sidebar The resizability is problematic, allowing the user to make the sidebar open, but with 0 width. It also uses the implicitWidth, which can cause it to jump around. The size is not remembered at all.
To address this, we can just declare a sensible width (minimum of GridUnit * 14 and root.width/2), and re-use the same logic used for the mobile actions and desktop thumbnails toolbars.

The loss of a user configurable size is unfortunate, but probably not fatal. We can't use SplitView for that because we want to animate it going off-screen rather than changing width (which would look awkward) and the current approach cannot feasibly be made to work without breaking.