ImageViewPage: Fix black frame when switching images
When switching with the arrow keys, there is (usually) a single black frame shown as the image is being loaded.

To fix this, the loader responsible for loading the images will now load when the image is the previous and next image, not just when it is the current image and the user is dragging.

This also helps resolve a short delay when dragging, for the adjacent images to become visible.

BUG: 506239