Make sure FileSystemTracker actually runs in a thread
FileTracker::setFolder is run on the main thread before it is moved. It is an expensive operation as KDirWatch has to iterate through all subdirs setting up notifiers.

After the tracker is moved to another thread and started calling methods directly on the Tracker are now incorrect behaviour. They will run on the main thread, which is unsafe. setSubFolder should be a posted message.

The thread cleanup was also unsafe. Calling QThread::quit only sends a message to that thread asking it to quit. We also need to wait until the thread has quit before deleting the thread.