BaseImageDelegate: Account for scaling such that images are shown at 100% scale pixel-to-pixel
For example, at a scale of 2x, to show the image such that one pixel in the image is equivalent to one pixel for display, we want sourceSize / 2.

Screen.devicePixelRatio is deficient for this, because it does not account for fractional scaling (outside of QT_SCALE_FACTOR). The window devicePixelRatio is correct, but not exposed to QML so it needs to be shimmed from C++.