VideoPlayer: Show the first frame of the video by preventing a stopped state
When the VideoPlayer is created and when playback finishes, it enters a stopped state which shows a black screen.

By pausing, we change to a state where the first frame of the video is being shown rather than a black screen.

Because we are never in a stopped state, the initial play button is not visible, and so this is removed (it also does not exist in Dragon Player, with which we want to be closer to).