RasterImageDelegate: Only stop being smooth at >= 400%
This matches Gwenview's behaviour: when the image is zoomed to 400% or greater, the image becomes pixellated.

The old behaviour was only to apply smooth & mipmapping when zoomed out (< 100%).

In order to prevent reloading the image (which causes a black frame), mipmap should not change. It has no effect unless zoomed out anyway.

retainWhileLoading would help, but causes noisy warning output. Either way, we should only have to load the image once, so masking reloading is not really ideal.

BUG: 503454