ThumbnailItem: Square thumbnails in most cases
KIO::PreviewJob returns a preview which fits within the request size, rather than fills. In order to fill, we can request a preview twice as large, and fill if we get an image that is larger or equal to the size of the preview.

When the image's aspect ratio is more extreme than 2:1 or 1:2, it will fit.