Fix mishandling of URL containing quotes in QUrl–QString dance
Rather than keep converting a path between QUrl and QString, just keep it as a QUrl and do the right things: we have a QUrl, ImageFolderModel wants a QUrl. Keep it as a QUrl.

We can instead use QUrl functionality to get the parent directory, rather than manipulate a string. This fixes an issue where 'Open With'/arguments to a file where a parent directory has quotes, literally "something", causes an error in ImageFolderModel.

BUG: 506465