thumbnailitem: Show mime type icon rather than loading icon
When the preview isn't ready, we currently show nothing and have a loading icon in QML.

This commit instead shows the mimetype icon instead of nothing, and removes the QML loading icon.

This is particularly good for folders: currently, a loading icon is shown for them momentarily, but we can show a representative icon immediately.