Customisable image view background & transparency checkerboarding
With this commit, the background color of the image view can be changed from black to white or the theme's default background color. Black is kept as the default.

Additionally, an image provider is implemented to provide a checkerboard pattern appropriate for the background color, which is used to show a checkerboard background behind images with transparency. Checkerboarding behind transparent images is enabled by default, but can be disabled.