EditorView: Implement 'Save As', improve back dialog and confirm saving changes
A new 'Save As' dialog is added so users can save their changes to a new image.

When saving changes, a dialog is shown to confirm this as the original image is irrevocably overwritten with the changes.

The back dialog now offers to save for you, rather than having you manually save after cancelling going back.

Shortcuts have been added for the saving and undo/redo actions.