Extensive refactoring on models and related code
This commit performs extensive ground-up refactoring on models, eliminating dead code, unused roles, etc and providing a more well-defined interface for models.

In particular, models either derive from a base class which is non-navigable (e.g. the Favorites model) or navigable (e.g. the Folder model). This path is implemented as a QVariant, and an improved navigator component extracted from AlbumView supports paths based on QUrl or QStringList. A name for each path item is provided by the model.

The tag, time and location model now have a 'root' location which shows collections for each tag, time group or location group. This will be useful for further work that might show tags as a single entry in the sidebar, rather than every tag (which may be a lot of items for those with many tags).

The gallery view now supports sorting by arbitrary file properties rather than just name (such as date created) and reverse sorting. Items can be searched by name.

In order to reduce items jumping around, as models may take time to load content, filtering is performed after sorting. Currently items are still filtered as they arrive in the base model, which means that there is still some jumpiness. Further changes could resolve this by having, for instance, the folder model update with the listed items at once, though the performance impact of having to filter (by checking mimetype) for each item at once would be problematic. Some amount of asynchronicity is called for here, where the folder model reports all items at once and subsequently sorting, then filtering, could be performed in different threads and reported back. It is not clear how best to implement this. Filtering takes the most time to perform (and is skipped in certain models based on database content), and sorting takes less time (but still has to check file properties on disk).

Many QML files are also moved to a subdirectory depending on their purpose (either used in GalleryPage or MediaViewPage).