Add file actions to the More Actions menu
Converts FileMenu into a singleton that provides QActions to be converted into Kirigami Actions.

The hidden actions are now separated by whether or not they affect the UI or deal with the file.

The org.kde.koko.private QML module was removed because there are no more classes left to define it.

PrinterHelper::printFileFromUrl is now static so that it can be used easily in C++.