Remove usage of cacheBuffer in our views
This is unnecessary and could even harm performance (depending on your
system.) It was used in AlbumView and ThumbnailStrip. Also for some
reason displayMarginBeginning/displayMarginEnd was set for the thumbnail
strip but I don't know why, so I removed that too.

For example, on my 4K screen at 125% scaling the album view can create a
staggering >2k delegates with our current cacheBuffer sizing! To put
that into perspective, using the default cacheBuffer size results in
only ~900 (which is still a lot IMO but at least Qt is deciding that.)
At 31x19 rows and columns, only ~589 can be shown at once anyway which
makes a large amount of these delegates redundant.

They also take up significant memory - at least hundreds of megabytes -
and it also adds a few seconds to our initialization time. Naturally
this depends on the size of your screen and what you're viewing, it's
all relative though. I didn't notice scrolling performance improving or
getting worse when going back to the default cacheBuffer size.