ThumbnailItem: Replace QPainter usage with scene graph
This is supposed to be faster, but the end benefit seems negligible. We
are no longer rendering to an indirect 2D surface at least, so that's
good and the speed is (better) but about the same. There's not a large
benefit to memory usage either, since we keep QImages cached anyway.

There is some fringe benefits to this approach too, I was able to
simplify a bit of the code and the filtering looks better. We can also
toy around with other interesting QSG features like atlasing in the
future.