InfoDrawerSidebarBase: Clean up tag list appearance and make tagging easier
Removes the ComboBox and Button for adding tags.

Adds a SearchDialog that can search for tags, add tags and remove tags.
Add/remove tags by checking/unchecking the list items.
Pressing Enter adds the current text as a tag or removes the current text if it is already a tag.
Pressing Delete on a selected list item removes the associated tag.
Hopefully there are no major performance issues for users with thousands of tags.

The dialog is opened by a Chip like the mockup in issue #100, although the appearance is not exactly the same.