Add floating zoom bar
It's a toolbar with combobox and zoom in/out buttons.
The combobox is supposed to handle input as it is typed from any locale
supported by Qt. When the text is updated, it should preserve the cursor
position relative to the end of the string before the percent so that
numbers can keep being inserted in the same number position. It should
avoid updating the text while the user is typing, but immediately update
the text when the zoom is changed while the user is not typing.
It should also handle entering numbers in more than just the current
locale. It doesn't need to handle every locale, but the order is current
locale -> en_US locale (if not already en_US) -> C locale. This way users
don't need to be too mindful about changing their habits based on their
system settings. en_US was picked as one of the fallbacks because a lot
of South American, African and Asian locales have similar number formatting.