Remove rounding of implicit content size
At 125% and with certain images (for example one at an odd size like
2560 x 1377) the checkboard pattern shows for a half-pixel or so past
the actual image bounds. As far as the Items themselves, according to
GammaRay they both have the same height but looking at the scene graph
the checkboard is slightly higher by a fraction.

Removing the Math.round here while calculating the implicit content size
fixes the issue for me, but I don't know why it does.